backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.11
python-dotenv==1.0.0
stripe==7.0.0
Pillow>=10.2.0
django-allauth==0.54.0
drf-yasg==1.21.7
django-parler==2.3
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0

